Refactoring set, get and disable usage export bucket samples by putting them in separate files. Removal of compute_instances_verify_default_value region tag as we decided not to use it.
